feat:Food Crud Finalizado

+++ src/main/resources/db/migration/V2__Create_Food_Table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE food (
     food_id SERIAL  PRIMARY KEY,
     food_name VARCHAR(255) NOT NULL,
-    calories INT NOT NULL,
-    proteins INT NOT NULL,
-    carbs INT NOT NULL,
-    fibers INT NOT NULL,
-    fats INT NOT NULL
+    calories NUMERIC NOT NULL,
+    proteins NUMERIC NOT NULL,
+    carbs NUMERIC NOT NULL,
+    fibers NUMERIC NOT NULL,
+    fats NUMERIC NOT NULL
 );
